Made a mistake. Commands changed from ints to shorts.

+++ robot/src/sdp/group2/comms/Commands.java
@@ -2,7 +2,7 @@ package sdp.group2.comms;
 
 /**
  * Opcodes for commands.
- * Each command should be an int[] made up of 4 bytes.
+ * Each command should be an short[] made up of 4 bytes.
  * First byte - opcode
  * Remaining three bytes will be options for the commands
  * @author Michael Mair
@@ -11,23 +11,23 @@ package sdp.group2.comms;
  */
 public class Commands {
 	
-	public final static int DO_NOTHING = 0;
-	public final static int FORWARDS = 1;
-	public final static int BACKWARDS = 2;
-	public final static int ANGLEMOVE = 3;
-	public final static int ROTATE = 4;
-	public final static int STOP = 5;
-	public final static int KICK = 6;
-	public final static int STEER = 7;
-	public final static int OPENKICKER = 8;
-	public final static int CLOSEKICKER = 9;
-	public final static int ROTATEKICKER = 10;
+	public final static short DO_NOTHING = 0;
+	public final static short FORWARDS = 1;
+	public final static short BACKWARDS = 2;
+	public final static short ANGLEMOVE = 3;
+	public final static short ROTATE = 4;
+	public final static short STOP = 5;
+	public final static short KICK = 6;
+	public final static short STEER = 7;
+	public final static short OPENKICKER = 8;
+	public final static short CLOSEKICKER = 9;
+	public final static short ROTATEKICKER = 10;
 	
-	public final static int FORCEQUIT = 63;
-	public final static int DISCONNECT = 64;
+	public final static short FORCEQUIT = 63;
+	public final static short DISCONNECT = 64;
 	
 
-	public static String getInitial(int command) {
+	public static String getInitial(short command) {
 		switch (command) {
 		case DO_NOTHING: 	return "STAY";
 		case FORWARDS: 		return "FW";
